refactor(mcp): extract idempotency manager
