+++ tldw_Server_API/app/core/MCP_unified/tool_execution/__init__.py
@@ -2,10 +2,12 @@
 
 from .coordinator import ToolExecutionCoordinator
 from .dependencies import CompatibilityCallbackLedgerEntry, ToolExecutionDependencies
+from .idempotency import IdempotencyManager
 from .reporting import ToolExecutionReporter
 
 __all__ = [
     "CompatibilityCallbackLedgerEntry",
+    "IdempotencyManager",
     "ToolExecutionCoordinator",
     "ToolExecutionDependencies",
     "ToolExecutionReporter",
